requirements.txt
Flask>=2.0
Flask-PyMongo>=2.3
Flask-Login>=0.5
Flask-Bcrypt>=1.0
python-dotenv>=0.19
Werkzeug>=2.0 # Ensure compatibility
pymongo[srv]>=4.0
email-validator>=1.1 # For basic email checks
click>=8.0 # For Flask CLI commands